+++ kubejs/server_scripts/Mods/Greate/Mixer.js
@@ -3,14 +3,4 @@ ServerEvents.recipes(event => {
 		'kubejs:unf_turtle_helmet',
 		Fluid.of('create:potion', 576, {Bottle:'REGULAR', Potion:'minecraft:water_breathing'})
 	]).heated()
-
-	event.recipes.greate.mixing('create:rose_quartz', [
-		'4x minecraft:redstone',
-		'minecraft:quartz'
-	])
-
-	event.recipes.greate.mixing(Fluid.of('gtceu:rubber', 144), [
-		Item.of('gtceu:raw_rubber_dust', 3),
-		Item.of('gtceu:sulfur_dust', 1)
-	]).heated()
 })